requirements.txt
py-clob-client>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
websockets>=12.0
aiohttp>=3.9.0
web3>=6.0.0
